feat: add src/hash function, move files helpers
