+++ src/fs/read.js
@@ -1,6 +1,6 @@
-import { getPathFromFiles } from "./getPath.js";
+import { getPathFromFiles } from "../getPath.js";
 import { readFile } from "node:fs/promises";
-import { errorMsg } from "./getErrorMsg.js";
+import { errorMsg } from "../getErrorMsg.js";
 
 const read = async () => {
   const fileToRead = getPathFromFiles(import.meta.url, "fileToRead.txt");
