+++ src/getPath.js
